``` Refactor: Extract bootstrap logic into dedicated package
- Moved application wiring from `cmd/api/main.go` to
  `internal/platform/bootstrap/`.
- Introduced `App` struct to encapsulate initialized components.
- Simplified `main.go` to delegate initialization to bootstrap package.
- Updated documentation to reflect the new structure.
  ```
