+++ .specify/feature.json
@@ -1,3 +1,3 @@
 {
-  "feature_directory": "specs/034-frontend-quality-improvements"
+  "feature_directory": "specs/035-bootstrap-extraction"
 }
